Updated project: added new files and modified existing ones

+++ src/main/java/com/project/uber/uberweb/dto/RatingDto.java
@@ -3,8 +3,10 @@ package com.project.uber.uberweb.dto;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class RatingDto {
     private Long rideId;
-    private Integer rating;
+    private BigDecimal rating;
 }
