+++ src/main/java/com/project/uber/uberweb/dto/RiderDto.java
@@ -3,6 +3,8 @@ package com.project.uber.uberweb.dto;
 
 import lombok.*;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
@@ -10,5 +12,5 @@ public class RiderDto {
 
     private Long id;
     private UserDto user;
-    private Double rating;
+    private BigDecimal rating;
 }
