+++ src/main/java/com/project/uber/uberweb/services/RatingService.java
@@ -4,11 +4,13 @@ import com.project.uber.uberweb.dto.DriverDto;
 import com.project.uber.uberweb.dto.RiderDto;
 import com.project.uber.uberweb.entities.Ride;
 
+import java.math.BigDecimal;
+
 public interface RatingService {
 
-    DriverDto rateDriver(Ride ride, Integer rating);
+    DriverDto rateDriver(Ride ride, BigDecimal rating);
 
-    RiderDto rateRider(Ride ride, Integer rating);
+    RiderDto rateRider(Ride ride, BigDecimal rating);
 
     void createNewRating(Ride ride);
 }
